+++ frontend/src/pages/components/Navbar.js
@@ -8,8 +8,8 @@ import {
 import { useState, useEffect } from "react";
 import { Link, useLocation } from "react-router-dom";
 import NavLinks from "./NavLinks";
-import { useAuthContext } from "../hooks/useAuthContext";
-import useGetUser from "../hooks/useGetUser";
+import { useAuthContext } from "../../hooks/useAuthContext";
+import useGetUser from "../../hooks/useGetUser";
 
 const Navbar = () => {
   const { user, dispatch } = useAuthContext();
