+++ frontend/src/pages/users/components/FooterSection.js
@@ -1,0 +1,17 @@
+import Pagination from "../../components/Pagination";
+
+const FooterSection = ({search, handlePageClick, pageCount, setItemsPerPage}) => {
+  return (
+    <div className="p-3">
+      {!search.length > 0 && (
+        <Pagination
+          handlePageClick={handlePageClick}
+          pageCount={pageCount}
+          setItemsPerPage={setItemsPerPage}
+        />
+      )}
+    </div>
+  );
+};
+
+export default FooterSection;
